Move integration tests for MySQL and PostgreSQL
Use SmokeTest because these run as part of every build.
