+++ presto-postgresql/src/test/java/com/facebook/presto/plugin/postgresql/TestPostgreSqlIntegrationSmokeTest.java
@@ -14,17 +14,23 @@
 package com.facebook.presto.plugin.postgresql;
 
 import com.facebook.presto.tests.AbstractTestIntegrationSmokeTest;
+import com.facebook.presto.tests.datatype.CreateAndInsertDataSetup;
+import com.facebook.presto.tests.datatype.CreateAsSelectDataSetup;
+import com.facebook.presto.tests.datatype.DataSetup;
+import com.facebook.presto.tests.datatype.DataType;
+import com.facebook.presto.tests.datatype.DataTypeTest;
+import com.facebook.presto.tests.sql.JdbcSqlExecutor;
+import com.facebook.presto.tests.sql.PrestoSqlExecutor;
 import io.airlift.testing.postgresql.TestingPostgreSqlServer;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.Test;
 
 import java.io.IOException;
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
-import java.sql.Statement;
+import java.util.function.Function;
 
+import static com.facebook.presto.tests.datatype.DataType.varcharDataType;
 import static io.airlift.tpch.TpchTable.ORDERS;
+import static org.testng.Assert.assertFalse;
 import static org.testng.Assert.assertTrue;
 
 @Test
@@ -55,6 +61,16 @@ public class TestPostgreSqlIntegrationSmokeTest
     }
 
     @Test
+    public void testDropTable()
+    {
+        assertUpdate("CREATE TABLE test_drop AS SELECT 123 x", 1);
+        assertTrue(getQueryRunner().tableExists(getSession(), "test_drop"));
+
+        assertUpdate("DROP TABLE test_drop");
+        assertFalse(getQueryRunner().tableExists(getSession(), "test_drop"));
+    }
+
+    @Test
     public void testInsert()
             throws Exception
     {
@@ -62,6 +78,16 @@ public class TestPostgreSqlIntegrationSmokeTest
         assertUpdate("INSERT INTO test_insert VALUES (123, 'test')", 1);
         assertQuery("SELECT * FROM test_insert", "SELECT 123 x, 'test' y");
         assertUpdate("DROP TABLE test_insert");
+    }
+
+    @Test
+    public void testViews()
+            throws Exception
+    {
+        execute("CREATE OR REPLACE VIEW tpch.test_view AS SELECT * FROM tpch.orders");
+        assertTrue(getQueryRunner().tableExists(getSession(), "test_view"));
+        assertQuery("SELECT orderkey FROM test_view", "SELECT orderkey FROM orders");
+        execute("DROP VIEW IF EXISTS tpch.test_view");
     }
 
     @Test
@@ -86,12 +112,87 @@ public class TestPostgreSqlIntegrationSmokeTest
         execute("DROP SERVER devnull");
     }
 
-    private void execute(String sql)
-            throws SQLException
+    @Test
+    public void testPrestoCreatedParameterizedVarchar()
     {
-        try (Connection connection = DriverManager.getConnection(postgreSqlServer.getJdbcUrl());
-                Statement statement = connection.createStatement()) {
-            statement.execute(sql);
-        }
+        varcharDataTypeTest().execute(getQueryRunner(), prestoCreateAsSelect("presto_test_parameterized_varchar"));
+    }
+
+    @Test
+    public void testPostgreSqlCreatedParameterizedVarchar()
+    {
+        varcharDataTypeTest().execute(getQueryRunner(), postgresCreateAndInsert("tpch.postgresql_test_parameterized_varchar"));
+    }
+
+    private static DataTypeTest varcharDataTypeTest()
+    {
+        return DataTypeTest.create()
+                .addRoundTrip(varcharDataType(10), "text_a")
+                .addRoundTrip(varcharDataType(255), "text_b")
+                .addRoundTrip(varcharDataType(65535), "text_d")
+                .addRoundTrip(varcharDataType(10485760), "text_f")
+                .addRoundTrip(varcharDataType(), "unbounded");
+    }
+
+    @Test
+    public void testPrestoCreatedParameterizedVarcharUnicode()
+    {
+        unicodeVarcharDateTypeTest().execute(getQueryRunner(), prestoCreateAsSelect("postgresql_test_parameterized_varchar_unicode"));
+    }
+
+    @Test
+    public void testPostgreSqlCreatedParameterizedVarcharUnicode()
+    {
+        unicodeVarcharDateTypeTest().execute(getQueryRunner(), postgresCreateAndInsert("tpch.postgresql_test_parameterized_varchar_unicode"));
+    }
+
+    @Test
+    public void testPrestoCreatedParameterizedCharUnicode()
+    {
+        unicodeDataTypeTest(DataType::charDataType).execute(getQueryRunner(), prestoCreateAsSelect("postgresql_test_parameterized_char_unicode"));
+    }
+
+    @Test
+    public void testPostgreSqlCreatedParameterizedCharUnicode()
+    {
+        unicodeDataTypeTest(DataType::charDataType).execute(getQueryRunner(), postgresCreateAndInsert("tpch.postgresql_test_parameterized_char_unicode"));
+    }
+
+    private static DataTypeTest unicodeVarcharDateTypeTest()
+    {
+        return unicodeDataTypeTest(DataType::varcharDataType)
+                .addRoundTrip(varcharDataType(), "\u041d\u0443, \u043f\u043e\u0433\u043e\u0434\u0438!");
+    }
+
+    private static DataTypeTest unicodeDataTypeTest(Function<Integer, DataType<String>> dataTypeFactory)
+    {
+        String sampleUnicodeText = "\u653b\u6bbb\u6a5f\u52d5\u968a";
+        String sampleFourByteUnicodeCharacter = "\uD83D\uDE02";
+
+        return DataTypeTest.create()
+                .addRoundTrip(dataTypeFactory.apply(sampleUnicodeText.length()), sampleUnicodeText)
+                .addRoundTrip(dataTypeFactory.apply(32), sampleUnicodeText)
+                .addRoundTrip(dataTypeFactory.apply(20000), sampleUnicodeText)
+                .addRoundTrip(dataTypeFactory.apply(1), sampleFourByteUnicodeCharacter);
+    }
+
+    private DataSetup prestoCreateAsSelect(String tableNamePrefix)
+    {
+        return new CreateAsSelectDataSetup(new PrestoSqlExecutor(getQueryRunner()), tableNamePrefix);
+    }
+
+    private DataSetup postgresCreateAndInsert(String tableNamePrefix)
+    {
+        return new CreateAndInsertDataSetup(postgresExecutor(), tableNamePrefix);
+    }
+
+    private void execute(String sql)
+    {
+        postgresExecutor().execute(sql);
+    }
+
+    private JdbcSqlExecutor postgresExecutor()
+    {
+        return new JdbcSqlExecutor(postgreSqlServer.getJdbcUrl());
     }
 }
